REFACTOR : Update code from suggestions
chore(testutil) : add testutil file for mock methods used in test files
chore(lint) : fix some lint issues
